src/Domain/Availability.java
/*
This file consists of Domain Availability class which implements the getters and setters.

As class members are private we can access the variables only through getters and setters.
 */
package Domain;

public class Availability {
    private int Flight_Number,Total_No_of_Seats,Avaliable_No_of_Seats;

    public int getFlight_Number() {
        return Flight_Number;
    }

    public void setFlight_Number(int flight_Number) {
        Flight_Number = flight_Number;
    }

    public int getTotal_No_of_Seats() {
        return Total_No_of_Seats;
    }

    public void setTotal_No_of_Seats(int total_No_of_Seats) {
        Total_No_of_Seats = total_No_of_Seats;
    }

    public int getAvaliable_No_of_Seats() {
        return Avaliable_No_of_Seats;
    }

    public void setAvaliable_No_of_Seats(int avaliable_No_of_Seats) {
        Avaliable_No_of_Seats = avaliable_No_of_Seats;
    }

}
